Get the manifest files from ManifestLoader
