+++ src/feature/controls_panel/views/control-component/ControlComponentState.ts
@@ -1,5 +1,5 @@
 import usePlotLayout from "@src/feature/common/usePlotLayout";
-import { FUNCTION_PARAMETERS } from "@src/feature/flow_chart_panel/manifest/PARAMETERS_MANIFEST";
+import { FUNCTION_PARAMETERS } from "@src/utils/ManifestLoader";
 import { ElementsData } from "@src/feature/flow_chart_panel/types/CustomNodeProps";
 import { ResultIO } from "@src/feature/results_panel/types/ResultsType";
 import {
